add countsort

+++ src/quicksort.h
@@ -1,7 +1,7 @@
 #ifndef QUICKSORT_H
 #define QUICKSORT_H
 
-void quicksort(int *A, int p, int r);
-void randomizedQuicksort(int *A, int p, int r);
+void quicksort(long *A, int p, int r);
+void randomizedQuicksort(long *A, int p, int r);
 
 #endif /* QUICKSORT_H */
